40/n: ssr: match parens
Summary:
Explicitly match parens that appear in the pattern.

Note that as SSR operates at the moment, the intial traversal over a file to identify match starting points always sees parens.

It makes sense to expose the parens strategy to the match, and use the same for both the fold and the match.

See following diff, D68555952

Reviewed By: robertoaloi

Differential Revision: D68555612

fbshipit-source-id: f9438cd10557daa002bbc2a70db25c6ded4b5904