BE: include the ExprId and PatId when tree printing an HIR tree
Summary:
When printing the HIR tree for a `Pat` or `Expr`, include the relevenat `PatId`/`ExprId` in the output.

This helps when debugging SSR matches, to compare the child `SubId` values to the original trees.

It is subject to a control variable in the printer, but this is hard coded for now.

Reviewed By: jcpetruzza

Differential Revision: D68706448

fbshipit-source-id: b0785b0b85db3a492bc45dfaa7a973644c7cbde0