ssr: store matches by placeholder `SubId`
Summary:
There can be more than one placeholder with the same underlying `hir::Var` as a name.

We need to ensure that where there are multiple placeholders for the same variable, they match the same code fragment.

As a step towards that we store all of the matches, using the placeholders unique `SubId` rather than the `hir::Var` as at present.

In following diffs we will check for equivalent values and fail the match if not.

Reviewed By: TD5

Differential Revision: D68775873

fbshipit-source-id: ef345eb1ddfb488ba7b7d768bf319e714aad4374