ssr: pass pattern SubId for code SubId into matcher
Summary:
We want to be able to re-use the matching machinery to check that two placeholder matches are the same.

This means that instead of comparing an HIR tree from the pattern body with one in the code body, we want to compare two different parts of the code body.

To enable this, we pass in the top-level `SubId` to match, instead of asking the pattern rule to return the one corresponding to the top level code `SubId` being matched.

This is the first step of the matching process anyway, so it may as well happen outside the matcher.

Note: there is a lot of parameter passing all over the place, it feels like it needs a cleanup to put things into a struct earlier in the process.

Reviewed By: TD5

Differential Revision: D68776673

fbshipit-source-id: 262ac1077b9f876b9b822d70e5cafacfa30b6dbb