BE: ssr: introduce PatternIterator::Leaf
Summary:
If we successfully match a pattern `SubId` with a code one, we need to recurse onto their children.

We have two kinds of child node, one which represents a linear list of `SubId`s that must be matched in order, and one for maps, which has a prefix in order and a set of fields that can be matched in any order.

When we hit a node without children, we have been encoding it as an empty linear list.

This diff introduces a third constructor `Leaf` to make this explicit. It also simplifies the test for if a returned `PatternIterator` is a leaf.

Reviewed By: TD5

Differential Revision: D68829220

fbshipit-source-id: de124ae3cbb9dceaa3327bb1dfd0c245ed6b8ec6