Highlight all function clauses, not just the first one
Summary:
Add the ability to specify multiple "highlight ranges" for the same definition. This is to be able to highlight, for example, multiple clauses for the same function definition.

This is achieved by adding an extra field to the "navigation target". The logic for calculating the "focus range" has been moved into the `to_nav`and could be further refactored, but that's currently out of scope for this diff.

Reviewed By: alanz

Differential Revision: D69050776

fbshipit-source-id: eccfe5e5b8b5e13616976004b4c0f93ad9f8f0c7