ssr: improve matching for PatternMap iterator
Summary:
For SSR matching within a map, we use a `PatternMap` structure, which maps first the name of the map in order, then the fields, allowing them to be in any order.

The matching algorithm partitions the field matches into those the have keys that are placeholders, which can match any key in the target map, and those that are not which have a precise match.

It first matches the non-placeholders, then arbitrarily assigns the placeholder matches to the remaining fields in the target.

Unfortunately this last step had been left out, this diff adds it in.

Reviewed By: TD5

Differential Revision: D69058907

fbshipit-source-id: 377c2fd907547b5d7a522cecbe0ed22ce2423753