Add shebang form for escripts
Summary:
Escript files have a shebang on top.

At the moment we do not do anything special, and hope that the tree-sitter error recovery is good enough, which it is at the moment.

But that is a matter of luck, and it can change as the grammar evolves.

This diff adds a specific top-level form to recognise a shebang, and this allows predictable parsing of the rest.

Reviewed By: TD5

Differential Revision: D69125735

fbshipit-source-id: 47b6991395afecb9d3f4fd6bc6e4019cf4a08147