Add support for CodeActionResolve
Summary:
This adds a middleware for resolving code actions. The idea is to intercept `CodeAction` requests and, in case a user input is required, an input box is presented to the user to get additional input.

This is useful, for example, in case of a "extract function" or "extract variable" operation, where the user can specify the name of the new function/variable without having to perform a subsequent rename.

Built-in validation is added for atoms and variables

Fixes #76.

Reviewed By: alanz

Differential Revision: D69112486

fbshipit-source-id: 6384c0cce1ff03d350abdc1630ce75b16c3e9bf5