Limit the max size of logged eqwalizer errors
Summary:
When people turn on debug logging, it can generate strings that are too long for VS Code to handle, > 512MB.

These typically originate from eqwalizer, when it returns a problem.

Add function to limit a string length for logging, and apply it to possible origins in Eqwalizer ELP code.

There are two variants of the limiting function, one takes a max length, the other uses a constant value.

Reviewed By: VLanvin

Differential Revision: D69311050

fbshipit-source-id: 53188669f2096f25a5e5809a7e49cdc807cdc96f