filter common_test::exportable_test_ranges by file_id
Summary:
We are seeing panics from ranges beyond the end of a file from the `exportable_test_ranges`. This can only be if the definition in question does not originate in the current file.

So filter out ones not in the current file.

This allows us to remove `clamp_range` introduced in D69457800

Reviewed By: robertoaloi

Differential Revision: D69467231

fbshipit-source-id: 31f6f4eff15e2bb25f7bf6c8d793598c0ecf3e9c