Restore "Analyze Stacktrace" functionality
Summary:
# Context

The "Analyze Stacktrace" functionality was broken for a long time. A user attempted to use it and reported it was not able to extract a stacktrace.

# Solution

This diff restore the basic functionality by:

* Refactoring some private code so it's testable
* Adding a unit test for the stacktrace extraction
* Updating the regex that was used to identify stacktraces in the terminal
* Adding an E2E tests
* Making the opening mechanism more solid

Despite of this, the functionality is quite fragile. I am considering if we should drop it altogether, instead.

Reviewed By: alanz

Differential Revision: D69518411

fbshipit-source-id: d6270e1e41ae357f6cea4f18f277b35d255bf2d6