Mitigate race causing ETXTBSY on eqWAlizer exe
Summary:
A race condition can occur between a thread trying to use the eqWAlizer executable, and the main thread creating the executable, leading to a panic with "text file busy".

See https://github.com/rust-lang/rust/issues/114554 for more details.

This mitigates the issue by attempting to start the eqWAlizer command in a loop, retrying if it fails with ETXTBSY.

Reviewed By: robertoaloi, TD5

Differential Revision: D69775408

fbshipit-source-id: 9f08258bdb762edbe9c75ed813b52cbb0fb303df