1/2: Based integers
Summary:
Erlang allows integer literals of the form `16#2F`, `2#1011` to represent numbers in alternate bases, where the base prefix can be 2 to 36.

Our grammar accepts these, but we lower them into HIR as `Missing` due to trying to parse them as a bare integer.

This diff introduces an explicit representation of a `BasedInteger`, storing the base and the value after conversion to base 10.

The actual conversion is done in the next diff, this just does the adjustment to the new representation.

Reviewed By: TD5

Differential Revision: D69785481

fbshipit-source-id: 519df7ca3aedb3bc53d2ef4d5e84d644a3f361c9