bxl: include new app_name attribute
Summary:
The buck2 Erlang prelude has been updated so that an application can have an optional `app_name` field, which is populated if the target name would clash with another one in the project if used as an Erlang application name.

Include this field in the result of invoking `elp.bxl`, and in the buck project model processing, use it in preference to the target name if present for the application name.

Reviewed By: TheGeorge

Differential Revision: D69842648

fbshipit-source-id: 09f11fbccb85c593de2f274c9651dad0b637a1a1