Deduplicate file watchers sent to client
Summary:
With the new buck2 bxl approach we have a lot more applications, and files in disparate overlapping directories.  This means the way we have been constucting the file watch lists, which assumed disjoint directories per application, is no longer valid.

This diff introduce a set to de-duplicate the list of watches.

Reviewed By: jcpetruzza

Differential Revision: D69914621

fbshipit-source-id: a4af956e307f34a4564d7882b604b7ed9534ba10