Server: trim DidSaveTextDocument registrations
Summary:
When a file is saved, it changes on disk, and the change is picked up by the client file watcher.  We configure the watcher to look for source files (`.erl`, `.hrl`) but also various configuration files.  We currently list the configuration files both in the registration for `DidSaveTextDocument` and the file watcher.  This is redundant, and means changes to the kinds of config files watched must happen in two places.

This diff reduces it to the file watcher only, which also means we do not get redundant notifications for them at run time, from both the `save` and `watch` notifications.

Reviewed By: robertoaloi

Differential Revision: D69921369

fbshipit-source-id: 63db952e11d38ef6b26a3fcbc366be10b7def3a1