Server: Do not double-process changes
Summary:
In ELP as an LSP server, we get notifications of changes in the project, from the client (e.g. VS Code).  For as-you-type changes, we simply update the store and process it once through the main loop.

For changes coming from the `DidSaveTextDocument` and `DidChangeWatchedFiles`, we call the same function to process them.

If the change requires a project model refresh, we can potentially end up with more than one running at the same time, if we blindly process both messages.

This diff adds a flag to the processing to indicate if it came from `DidSave` or `Watch`, and ensures that processing a given file only happens once.

Reviewed By: robertoaloi

Differential Revision: D69921658

fbshipit-source-id: d6447c3f68ab4b066c8e2812dcd64a075a9f6bb4