Server: do not switch projects if pending config changes
Summary:
Project reloading has two stages.  The first is to generate the project model configuration from the config files, and the second is to put it into place, by populating internal data structures and loading the source files. This separation allows a common second stage, and multiple possible project definition options.

Now that we have a `ReloadManager` in the server state, use it to gate whether to do the second stage or not. This is done by checking if there are any pending changes to config files that would trigger a reload, and deferring to them if so.

A future diff could contemplate passing this info into the reload task earlier to abort it.

Reviewed By: robertoaloi

Differential Revision: D69994455

fbshipit-source-id: e6a34f44af121fc54180f751be10cb4bc4e5e527