Introduce a wait before reloading a project
Summary:
We receive notifications from the client when files change on disk. We watch for project configuration files, like `BUCK`, and reload the project if any of these change.  But if a branch switch happens, many of these could change in a short time. We can not be sure that all the changes will arrive in one batch, and so we want to avoid triggering project reloading, an expensive operation, if there are more imminent changes due.

This diff handles the scenario by inserting the changed config files into `ReloadManager`, which notes the time the most recent file was inserted. In the main loop we check that sufficient time has elapsed for the batch to have ended, and if so launch project reloading.

The initial wait time is 500ms, as a balance between ensuring we wait for the last change to come, and not delaying the reload too long.  The reload itself can be in the order of 10s, so the delay does not add appreciably to is, and prevents a second one straight after the first.

Reviewed By: robertoaloi

Differential Revision: D69983249

fbshipit-source-id: c70505ffb74c657c349e5e27e6f2cf9c20f09534