Better progress reporting for buck project loading
Summary:
The current progress reporting is not great, it says we are loading a project, finishes, then after a time the files start loading.

This diff provides progress reporting all the way through, including the various phases of buck project loading.

This is particulary important with the new BXL query approach, as the query may take some time on a branch switch, because BUCK may need to gather more information, and in future do a partial build.

Also, to assist in end-to-end tests, include the progress item title in the end message, so we can wait for the end of file loading.

Reviewed By: robertoaloi

Differential Revision: D69924285

fbshipit-source-id: 5160bc12353bd8040154f8c8d39b6d9b469f9e5c