ssr: query a match for comments
Summary:
We use an ssr match as the basis for constructing assists.

When modifying code it is important not to discard comments.

Facilitate this by providing methods on the match as  whole and each `PlaceHolderMatch` to retrieve any comments contained within it.

This information can be used to either decide not to generate an assist, or to construct one that preserves the comments

Reviewed By: TD5

Differential Revision: D70165132

fbshipit-source-id: 4d967f714f43a8deb27d5806eac6f874a08794b6