BE: server: remove Status::Invalid
Summary:
Our server status currently has a status of `Invalid`, which is entered if we happen to load a project with zero files.

This is not particularly useful, so this diff instead sends a pop-up message to the user to check their project configuration, and transitions to `Running`.

When the project config files change, we will notice and load the project.

Reviewed By: robertoaloi

Differential Revision: D70172702

fbshipit-source-id: dfd75cd64acfd6e3a67b831111946bf2d7a54d12