Move annotations extraction to project_model crate
Summary:
# Context

We want to introduce a new diagnostic which can warn about and auto-fix usages of old-style EDoc comments.

# Problem

The current testing framework used in ELP to write declarative "test snippets" has the concept of "annotations", which can be used to set expectations when computing diagnostics. They typically look like this:

```
%%  ^^^^^^^^^^^^^ 💡 warning: Function is undefined.
```
Currently, annotations are kept in the snippet as regular comments. This poses a problem, since it makes it hard to specify a diagnostic for an EDoc comment, since the annotation itself would be considered as part of the EDoc comment itself.

# Solution

We want to strip the annotations from the test snippet before computing diagnostics. To do so, we will move the extraction during the parsing stage. This is similar to what we already do to extract "tags" (e.g. for folding ranges), which would otherwise result in syntax errors.

# This Change

This diff mechanically moves the extraction logic from the `ide` to the `project_model` crate, in preparation for moving the extraction to the parsing stage. By having a mechanical move here, we will simplify the review process later, since the extraction logic will require some modifications.

Notice that a better place for the parsing logic would be the `base_db` crate, but we are trying to minimize the changes. We can consider moving there at the end of the stack.

Reviewed By: alanz

Differential Revision: D70165753

fbshipit-source-id: 1d3ef1c3d80c5c6d6023c7d2e34400f374d49ac4