Return text without annotations in extract_annotations function
Summary:
# Context

Here we extract annotations during parsing and we store the *sorted* annotations in the fixture. We update the `extract_annotations` function to also return the text without annotations.

Both the extracted annotations and the text without annotations are still unused at this stage. This will change next.

# Problem

A couple of changes are needed to the extraction logic. It is possible for a test fixture to contain:

```
//- /erl/my_app/src/main.erl
~%% <<< 💡 error: Some
```

Where the cursor sits before a "top_of_file" annotation.

# Solution

Until now, we were extracting annotations *after* the cursor marker, but this is not the case any longer. So, we need to take the marker into account. We do so in the extraction logic and in the regex to identify annotations.

An alternative would have been to keep extracting the marker before extracting the annotations. Unfortunately, that would be problematic since later in the stack we want annotation ranges not to include the annotations and markers themselves. So a proper fix would require a complete rewrite of the extraction logic. For now, this will do.

Reviewed By: alanz

Differential Revision: D70171225

fbshipit-source-id: 4e7ea1a4df12f88faf04b08c0f39e58dda1a3246