Edoc native diagnostics
Summary:
Now that we can finally specify a diagnostic covering a comment in the test fixture, let's start with implementing EDoc diagnostics :)

# Context
We want to auto-migrate Erlang documentation to the new Markdown format.

# Problem
The current solution provided  by the Erlang/OTP team is [currently broken](https://github.com/erlang/otp/issues/9031) and it is not a priority for the team to maintain it, given the deprecation of EDoc. Even if we were to fix support for it, the script does not actually convert EDoc documentation, but simply adds a copy of the documentation in the new format, leaving the old documentation intact. This is not ideal, since we would like a full migration.

# Solution
We already have support for extracting EDoc comments in ELP. We can leverage that to create a diagnostic that can show a warning when the old-style syntax is used. We can then provide an auto-fix associated to the diagnostic, which can perform the conversion automatically.
This has a few advantages:
* The linter can be run as an ELP CodeMod via the CodeMod service, which only requires a small configuration
* We prevent introducing new regressions since the linter would trigger in case developers introduce a new comment using the old syntax
* This can also help adoption of ELP in the open-source community, since it could just convert the docs with a single command

The current EDoc extractions has some limitations, which will need to be addressed before we can roll out the diagnostic.

# This change

Here we introduce a new diagnostic (`W0038`, aka `old_edoc_syntax`) that highlights EDoc comments using the old-style EDoc comments.
For now, we only introduce the basic plumbing and highlight the whole line, which will improve later in the stack.

The diagnostic is currently disabled by default.

Reviewed By: alanz

Differential Revision: D69936947

fbshipit-source-id: 2afa3b1d344b9030663c1cb7497ccc5ba7c9840c