Consume annotations from fixture instead of extracting them manually
Summary:
Expose the extracted annotations via the `ChangeFixture` data structure, so that callers can use them, instead of extracting manually.

We also tweak the annotation extraction logic to take into account the marker and the fact we're going to incrementally remove annotations from the text. Specifically, we:

* Fix the max file range to exclude annotations
* Distinguish between an annotation and a regular comment
* Adjust the line lengths to be `0` in case of annotations
* Adjust the line lengths to take cursors into account
* In the case a top-level cursor is used before an annotation, keep it

We also tweak the regex used to identify annotations not to always require a space after the range, since some of the tests define them that way.

An auxiliary `with_fixture` function is also added for convenience. There's some repetition between the fixture module in the ide crate and the fixture module in the base_db one. This should be addressed as a separate follow-up.

The glean tests are updated since they contain the explicit offset, which is now different since annotations are excluded from the final text.

We can probably avoid the double pass during the extraction, but that can be done as a follow up. Here we try to minimize the required changes.

Reviewed By: alanz

Differential Revision: D70171802

fbshipit-source-id: 6c1d4450728255ffc8020928c9f966b9a0e3ac84