Error details about partial union matches
Summary:
Add error details when the expression's type is a union where some candidates match the context but others do not.
As before, the chosen candidate is the one with the lowest score.

For now, this is only displayed in two situations: either it is a leaf node of the error message, or it is the root node (depth = 1). This is to limit verbosity, we may reconsider in the future, depending on feedback.

Reviewed By: ilya-klyuchnikov

Differential Revision: D70174753

fbshipit-source-id: 70b044739bfa67acc326ae4520ff6811b679e7db