maps:fold - special handling of lambdas with atomic keys
Summary:
Special handling of lambdas with ALL atomic keys in `maps:fold`.

For each key the most precise type of value can be given (for map types with atomic keys).

Pro: it's practical and it works.
Con: Occurrence typing is not applied (in this version), though it doesn't make it worse for our codebase.

Reviewed By: VLanvin

Differential Revision: D70251104

fbshipit-source-id: 7477f316c5de2db424a4da0b4fef31006368630c