1/n BE: tree_print body : RecordBody
Summary:
This sequence of diffs improves the discoverability of the `tree_print` debugging feature by providing the method on each of the kinds of `Body` that we lower.

Tree printing is a representation of the HIR AST, expanded as a tree, with the index values displayed.  This allows easier reasoning about e.g. diagnostic code when writing it.

Reviewed By: robertoaloi

Differential Revision: D70708284

fbshipit-source-id: a2f5955af7509806e9fc97975f30c58dcc3d9162