Convert HTML entities during conversion to Markdown
Summary:
# Context
We are implementing a quick-fix in ELP that can auto-convert old-style EDoc comments into Markdown.
# Problem
When authoring Erlang EDoc, developers have to explicity use HTML character entities to represent symbols which can be confused with tags (e.g. `&gt;` for `>`).
When converting to Markdown, we want HTML entities to be translated automatically.
# Solution
Vendor the `htmlentity` crate and use it to perform the conversion.

Reviewed By: jcpetruzza

Differential Revision: D70699249

fbshipit-source-id: bbda58cf00e2bd09e7b2a5d6102894c093713e80