Handle EEP59 params in signature help
Summary:
# Context
We are implementing a quick-fix in ELP that can auto-convert old-style EDoc comments into Markdown.
# Problem
The *signature help* functionality currently extracts parameters information from the `param` EDoc attributes. By moving to Markdown-style documentation, we lose this information.
# Solution
When populating the `def_map`, keep track of the last doc attribute encountered. Then, associate it to the function definition, so that the relevant information can be extracted.
Notice that, ideally, we would use a custom attribute to indicate the params. This is unfortunately not yet possible in Erlang/OTP (see [this discussion](https://github.com/erlang/otp/issues/9031)).
Also, in EEP 59 types and callbacks could also leverage the `-doc` attribute. Support for this will be added later.

Reviewed By: jcpetruzza

Differential Revision: D70704065

fbshipit-source-id: 80c92464a97400e4656805d04708a15b3b1d97fb