Do not lose @params information during conversion to Markdown
Summary:
# Context
We are implementing a quick-fix in ELP that can auto-convert old-style EDoc comments into Markdown.
# Problem
EDoc documentation can contain multiple occurrences of `param` to indicate parameter documentation, as in:
```
%% doc
%% Some doc
%%
%% param A One param
%% param B The other param
%% end
```
Right now we lose this information.
# Solution
Ensure parameter information is translated to Markdown. The EEP 59 does not specify how params information (which is used by signature help) is stored (see [this discussion](https://github.com/erlang/otp/issues/9031)), so we'd need to find a way to store it in a structured way.
Notice that (see https://github.com/erlang/otp/pull/5712) Erlang/OTP does not yet support custom attributes everywhere in the file, so this will have to be part of the plain-text doc for now.

Reviewed By: jcpetruzza

Differential Revision: D70704066

fbshipit-source-id: ab803ea36eee83c684aa137b623a33dabfaddb64