Auto-convert EDoc style inline doc-blocks
Summary:
# Context
We are implementing a quick-fix in ELP that can auto-convert old-style EDoc comments into Markdown.
# Problem
In EDoc, verbatim quoting is achieved by using a peculiar combination of backquotes and single quotes as described here.
Markdown uses backquotes consistently.
# Solution
Here we use a simplistic approach using a regex. Specifically, we always convert triple quotes into backquotes. While not necessarily correct, this should work in most (if not all) concrete cases. If not, we can tweak the regex or just edit the code. Not investing too much into the conversion since we will only need it once.

Reviewed By: jcpetruzza

Differential Revision: D70655184

fbshipit-source-id: 039189574511d45a53032279ab7844a235f93587