7/n: BE: FoldBody in tree_print
Summary:
A `Body` contains a number of arenas for storing the relevant parts of the HIR AST it supports, one each for expressions, patterns, type expressions and terms.

These are indexed via `ExprId`, `PatId`, `TypeExprId` and `TermId` respectively.  These indexes only work at on the specific arena variable, so we provide helper implementations of the `Index` trait on `Body` for each of these Id types, to do the indexed lookup on the appropriate arena variable.

When lowering to HIR, we explicitly capture the presence of additional parens in the original expressions, as well as the occurrence of macro calls.

When analysing the HIR AST these are irrelevant, so the `Body`-level indexing operations take care of "looking through" them, so if you provide the index of say a `MacroCall` it will return the HIR AST item for the expansion, not the macro call.

However, depending on the analysis being done, sometimes it is important to see the macro calls and/or parens.

This is made possibly by the introduction of a `FoldBody`, which contains a `Body` and the strategy to be followed when resolving the index for a `MacroCall` or Paren` item, to either look through it or return it.

Note that it is called a `FoldBody`, because the `Strategy` parameter passed to a `fold` call is used to configure the `FoldBody` used when walking the HIR AST for the fold.

# This Diff

This diff modifies the tree printing context to use a `FoldBody` instead of a `Body`, thus tailoring exactly what is shown.

The top-level custom body printers introduces earlier in the stack still use the default strategy of hiding macro calls and parens, but this can be changed later as needed.

Reviewed By: TD5

Differential Revision: D70710437

fbshipit-source-id: fbf8b4df27b4e40a62628248a6754873706dcf45