Fix Body indexing for nested macro calls and parens
Summary:
A `Body` provides indexing into the HIR AST, and is intended to provide an uncluttered view of it for semantic analysis.

But we capture macro calls and parens in this AST, because under some circumstances it is important to know where they are.  This is normally hidden, because the indexing mechanism skips straight to the expansion when it sees a macro call.

But there is a bug, in that it returns the expansion from the arena directly, which can be a nested macro call.

Fix this by updating the indexing operations to recursively call themselves, and so continue until the chain ends.

Reviewed By: TD5

Differential Revision: D70774014

fbshipit-source-id: 09e2634f1c2b969afeec9efce85b2101b7ca2a3f