make body indexing options more discoverable
Summary:
A `Body` captures an HIR AST in a series of arenas, one each for expressions, patterns, types and terms. These are indexed by `ExprId`, `PatId`, `TypeExprId` and `TermId` respectively.

When doing a fold, over the AST we can control visibility of macro calls and parens using a `Strategy`.  Under the hood, this constructs a `FoldBody` which controls the `Index` implementation for each of the `Id` types.

But a `FoldBody` is generally useful too, so this diff exposes a methof on `Body` to make this apparent.

Reviewed By: TD5

Differential Revision: D70773555

fbshipit-source-id: c4e1e0c3e94053b5327ca97fe9bb376cdbc4fb5b