macro precedence surprise diagnostic
Summary:
For the `erlc` compiler, a macro is expanded at the lexical level, so the parser never sees it.

At the surface level, when writing a macro call, it looks like a function call, and you have to look inside the macro to know what it expands to (or hover).

This can lead to surprises if the top level of the macro is a bare binary operation and it is used within another binary operation, as the precedence can then escape.  For example

```
-define(MYOP(B,C), B + C).
foo(A,B,C) -> A * ?MYOP(B,C).
```

will end up being parsed as `(A * B) + C`,. where at first glance it migh appear that it would be `A * (B +C)`.

This diff adds a diagnostic where this occurs, and offers to add parens at the macro call site, so leading to

```
foo(A,B,C) -> A * (?MYOP(B,C)).
```

Reviewed By: TD5

Differential Revision: D70775461

fbshipit-source-id: 47447e9424fc6e9033dd3a8874444a87f52b7035