Watch for changes to `.elp.toml` wherever they occur in the project
Summary:
For a buck project model, we set the project source root to be the root of the buck cell we are in.  But it may be that the `.elp.toml` config file is in a subdirectory, as is the case for common_test in the buck2 Erlang prelude.

Update the file watch spec to look for `.elp.toml` and `.elp_lint.toml` wherever they may occur in the project, so that they can trigger project reloading if needed.

Reviewed By: jcpetruzza

Differential Revision: D70969174

fbshipit-source-id: 58b2a9e15b8b7c3a3636e72a30cc758cc0a4abd0