Rename MakeDiag into Make
Summary:
# Context

The `find_call_in_function` API is tightly coupled to the concept of a diagnostic (e.g. it takes a `Vec<Diagnostic>` accumulator.

# Problem

This is not generic API, since the same folding could be reused in other context.

# Solution

In this stack we aim at making the API more generic. Here we simply rename `MakeDiag` into `Make`.

Reviewed By: jcpetruzza

Differential Revision: D70972626

fbshipit-source-id: 3c4778c10fa353cb798863b69e7037533f25c5b2