Make "Make" a parametrized type
Summary:
# Context

The `find_call_in_function` API is tightly coupled to the concept of a diagnostic (e.g. it takes a `Vec<Diagnostic>` accumulator.

# Problem

This is not generic API, since the same folding could be reused in other context.

# Solution

In this stack we aim at making the API more generic. Here we make the `Make` type parametrized.

Reviewed By: jcpetruzza

Differential Revision: D70972628

fbshipit-source-id: 24ce746a1c05e3d04fa8ad05b2c57e94714eb4ef