Rename MakeDiagCtx into MatchCtx
Summary:
# Context

The `find_call_in_function` API is tightly coupled to the concept of a diagnostic (e.g. it takes a `Vec<Diagnostic>` accumulator.

# Problem

This is not generic API, since the same folding could be reused in other context.

# Solution

In this stack we aim at making the API more generic. Here we simply rename the context not to refer to diagnostics. No functional changes.

Reviewed By: alanz

Differential Revision: D70972627

fbshipit-source-id: ef1a25692f8e58ec5f32fb3618f9a365f423a022