Parametrize find_call_in_function
Summary:
# Context

The `find_call_in_function` API is tightly coupled to the concept of a diagnostic (e.g. it takes a `Vec<Diagnostic>` accumulator.

# Problem

This is not generic API, since the same folding could be reused in other context.

# Solution

In this stack we aim at making the API more generic. Here we make the `find_call_in_function` API generic with regards to the type of the accumulator vector.

Reviewed By: jcpetruzza

Differential Revision: D70972625

fbshipit-source-id: c1e0de435b72c47609829e1b8ba25eee81eedf8b