bxl mode: Do not add included to applicable_files
Summary:
At some point we updated `elp.bxl` to return the include path directory, not a list of the individual files in it. This means we do not have to collapse it down in ELP, and reduces considerably the size of the generated json output.

But at the time we did not update the `applicable_files` field in the ELP `BuckTarget`, so we needlessly add entries to be processed as files, which are not.

Reviewed By: robertoaloi

Differential Revision: D71124093

fbshipit-source-id: eab88f2f80e8544dc68d16d42cf195e16161d3a1