bxl mode: Actually update the AppDataIndex in salsa after modifying it
Summary:
A buck target always explicitly lists the files it cares about. As a result, you can have multiple buck targets in the same directory, each listing different sets of files they care about. This happens particularly with tests, and a file should only belong to a single target.

When using bxl mode, we apply this distinction properly, and map one buck target to one ELP application. This means we can't rely on a directory, and hence `SourcRootId` to determine what the application is.

Instead, the `elp.bxl` query returns a list of `applicable_files` for each target, explicitly enumerating the files belonging to the target.  We use a two-level lookup process from a `FileId` to its `AppData`, first consulting a map of `FileId` to `AppDataId`, constructed from the `elp.bxl` query output, and then looking up the `AppData` using the `AppDataId`.

The project loading process goes through various steps
- project discovery
- construct a `Project`, in terms of absolute paths
- turn this into an appropriate internal representation
- load the files, assigning `FileId`s in the process.

The mapping from `AppDataId` to `AppData` is known in terms of absolute paths after the third step.  For CLI mode, we get to this point, load the files, assigning `FileId`s which we use to populates the `file_app_data_id` salsa input, then start operating.

For server mode it is not this simple, as we set up the project data, then activate an async VFS file loader.

So we keep a map of `AbsPathBuf -> AppDataId` in the server, and as each file is loaded and appears in the change processing, we set the appropriate input.

This happens via a call to `set_app_data_id_by_file()`. But, this function has a bug. It updates the contents of the stored data, but does not replace it in the salsa input, so the change is never made permanent.

# This diff
Fixes that, by doing the update.

Reviewed By: robertoaloi

Differential Revision: D71125404

fbshipit-source-id: cf4e61ad318c77e897b2478faf052503c80e030f