bxl mode: Backfill app_data_id's when project discovery is complete
Summary:
The previous diff ensures that we populate the necessary salsa inputs to be able to reliably retrieve the correct `AppData` for a given `FileId`.

But there is still one problem.  We only have the table of file paths to `AppDataId` after the initial project processing is done. But we start populating our salsa file store as soon as ELP launches,  based on files open at the time, or opened between launch and project discovery being complete.

# This diff
This diff introduces a process to sweep through the remaining entries that need to be populated after project loading is complete, and fill in the values.

Reviewed By: robertoaloi

Differential Revision: D71125906

fbshipit-source-id: e004a04b5c3a86e7a25d8b2b63f4e21449133a44