Operate EDoc extraction at hir level
Summary:
In preparation for the bug fix, we want to operate at hir level as much as possible. This will make it simpler, later in the stack, to check whether a function definition has a corresponding `-spec` attribute.

We take the opportunity to split extraction of module/function docs in separate functions.

Reviewed By: alanz

Differential Revision: D71189274

fbshipit-source-id: 763a9e5cf35888d5ba913f3c47608be1a59a480a