More refined module index populating
Summary:
Do not insert a file in the module index if it is an Erlang configuration file, or is in a SUITE data directory.

This matters with the new buck bxl-based project model, where we load a wider range of directories, some of which include assorted config and test files.

Reviewed By: robertoaloi

Differential Revision: D71203400

fbshipit-source-id: 891d9966408b6e65501b442be6bc887caceab7f5