More refined module index populating, take 2
Summary:
D71203400 made sure that no erlang configuration files entered the module index by discarding any that the tree-sitter-erlang parser recognised as configuration files.

But this meant that populating the initial module index required every file to be parsed by the tree-sitter-erlang parser, which dramatically slowed down what should be a simple book-keeping phase.

We already have a more blunt mechanism to exclude files, using a regular expression consulted when determining the `FileKind`, based on the file path only.

# This diff

Restores the effect of the reverted diff, but based on `FileKind`, which excludes ignored files when determining `is_module()`

Reviewed By: michalmuskala

Differential Revision: D71386533

fbshipit-source-id: 7d9756044c85ec4460db173faf7dfef9a51c0327