buck bxl: mark files from the prelude as third-party
Summary:
The buck2 BXL-based project model is able to track dependencies into the buck2 erlang prelude.  This is needed to ensure we are able to resolve references, types and so on. But they are not (normally) part of the project being worked on, and the current project does not have the power to change them.

# This diff

Marks targets originating in the prelude as `TargetType::ThirdParty` so they are not processes as part of the current project.  In particular, features like `parse-all`, `eqwalize-all` will not report diagnostics on them.

Reviewed By: TheGeorge

Differential Revision: D71397638

fbshipit-source-id: 97cf3806d85f35444114e92c811c71049b2167dd