Show @returns tag not being handled
Summary:
#Context
We are implementing a quick-fix in ELP that can auto-convert old-style EDoc comments into Markdown.
# Problem
In old-style EDoc, the `returns` tag can be used to add extra information about the return value of the function. We currently discard this tag.
# Solution
We should not discard the `returns` tag. Since there's no corresponding Markdown, we can use plain-text such as `Returns:`.

This diff simply highlights, via a new testcase, the problem which will be fixed later in the stack.

Reviewed By: Balajiganapathi

Differential Revision: D71313041

fbshipit-source-id: bdb059b96f7c496cd4c069a60d7fc2d251cb5d0e