eqwalizer: intern all strings in eqwalizer AST
Summary:
With this change we intern all strings in eqwalizer AST. This is done by switching from `SmolStr` to `Ustr`. I'm not sure how much of an impact interning everything has (given that SmolStr doesn't allocate for strings shorter than ~20 characters). However, we include those strings in many data structures and `SmolStr` is 24 bytes in size, while `ustr` string is only 8 bytes - this shrinks pretty much all data structures that eqwalizer uses.
In particular, things like `Id` are shrunk from 32 bytes to 16, and `RemoteId` from 56 to 24.

With some testing with `/usr/bin/time -v` this shrinks the max ELP size by ~1GB from ~8.85 to ~7.8, or ~11%.

The trade-off is that once allocated, those strings are never freed. However, the identifiers in AST are fairly stable, and I don't think this should be an issue in practice - we don't reload a lot of code for eqwalizer - we convert only once the code passes `erlc` linter, so it's mostly valid.

Why not salsa interning? It gives us back an ID we have to use back with salsa to get the value - this is API-incompatible with how serde works, so we wouldn't be able to "just" use it here.

Reviewed By: robertoaloi

Differential Revision: D71335139

fbshipit-source-id: 7c14d75d1eec07d298403e70c882311c7e75f50d