Fix deletion in case of intersperse documentation
Summary:
# Context
Old-style EDoc allows the concept of "intersperse" documentation, where attributes such as `-type` or `-record` can sit in between a `doc` tag and a function.
# Problem
Since we use the `cover` method to expand the EDoc header range every time we encounter a new comment, everything in between becomes (incorrectly) part of the header.
# Solution
Instead of storing a single range for the EDoc header, store a vector of ranges. This makes it possible to delete intersperse EDoc

Reviewed By: jcpetruzza

Differential Revision: D71316285

fbshipit-source-id: f4af61da224ac380d9e4601a455abd2822855cde