eqwalizer: reduce cloning in expander
Summary:
Before we'd clone the entire AST. Now we only clone the forms that end up in the stub - so we don't need to clone all the functions. This should be noticeable for memory/perf.
I also think the regular `for` ends up nicer here - we need less `Ok(())` decoration.

Reviewed By: ilya-klyuchnikov

Differential Revision: D71546366

fbshipit-source-id: de4277b0cdc3dc323259f657704c70e973911fe6