Take into account potential inline comments when inserting moduledoc
Summary:
It can happen that the "most significant attribute" after which we want to insert the module doc has an inline comment.

In that case, we want to preserve the location of the comment.

Add a utility functiom to "skip to the end of the line", then use it when inserting module docs.

Reviewed By: alanz

Differential Revision: D71613448

fbshipit-source-id: f664f3f224f1c74681ff6d00ee075583a21329da